close development and merge
